Refine routers, both api and html
Adjust models

+++ src/main/kotlin/github/buriedincode/bookshelf/routers/api/UserApiRouter.kt
@@ -2,43 +2,44 @@ package github.buriedincode.bookshelf.routers.api
 
 import github.buriedincode.bookshelf.Utils
 import github.buriedincode.bookshelf.models.Book
-import github.buriedincode.bookshelf.models.IdValue
 import github.buriedincode.bookshelf.models.ReadBook
 import github.buriedincode.bookshelf.models.User
 import github.buriedincode.bookshelf.models.UserInput
+import github.buriedincode.bookshelf.models.UserRole
 import github.buriedincode.bookshelf.tables.ReadBookTable
 import github.buriedincode.bookshelf.tables.UserTable
-import io.javalin.http.BadRequestResponse
 import io.javalin.http.ConflictResponse
 import io.javalin.http.Context
 import io.javalin.http.HttpStatus
 import io.javalin.http.NotFoundResponse
+import io.javalin.http.UnauthorizedResponse
 import io.javalin.http.bodyValidator
 import org.apache.logging.log4j.kotlin.Logging
 import org.jetbrains.exposed.sql.SizedCollection
 import org.jetbrains.exposed.sql.and
-import github.buriedincode.bookshelf.models.UserInput.ReadBook as ReadBookInput
 
-object UserApiRouter : Logging {
-    fun listEndpoint(ctx: Context): Unit =
+object UserApiRouter : BaseApiRouter<User>(entity = User), Logging {
+    override fun listEndpoint(ctx: Context) {
         Utils.query {
-            var users = User.all().toList()
-            val username = ctx.queryParam(key = "username")
-            if (username != null) {
-                users = users.filter {
+            var resources = User.all().toList()
+            ctx.queryParam("has-image")?.lowercase()?.toBooleanStrictOrNull()?.let { image ->
+                resources = resources.filter { it.image != null == image }
+            }
+            ctx.queryParam("username")?.let { username ->
+                resources = resources.filter {
                     it.username.contains(username, ignoreCase = true) || username.contains(it.username, ignoreCase = true)
                 }
             }
-            ctx.json(users.sorted().map { it.toJson() })
+            ctx.json(resources.sorted().map { it.toJson() })
         }
+    }
 
     private fun Context.getInput(): UserInput =
         this.bodyValidator<UserInput>()
-            .check({ it.role >= 0 }, error = "Role must be greater than or equal to 0")
             .check({ it.username.isNotBlank() }, error = "Username must not be empty")
             .get()
 
-    fun createEndpoint(ctx: Context): Unit =
+    override fun createEndpoint(ctx: Context) {
         Utils.query {
             val input = ctx.getInput()
             val exists = User.find {
@@ -48,7 +49,7 @@ object UserApiRouter : Logging {
                 throw ConflictResponse(message = "User already exists")
             }
             val user = User.new {
-                imageUrl = input.imageUrl
+                image = input.image
                 role = input.role
                 username = input.username
                 wishedBooks = SizedCollection(
@@ -72,22 +73,15 @@ object UserApiRouter : Logging {
 
             ctx.status(HttpStatus.CREATED).json(user.toJson(showAll = true))
         }
-
-    private fun Context.getResource(): User {
-        return this.pathParam("user-id").toLongOrNull()?.let {
-            User.findById(id = it) ?: throw NotFoundResponse(message = "Unable to find User: `$it`")
-        } ?: throw BadRequestResponse(message = "Unable to find User: `${this.pathParam("user-id")}`")
     }
 
-    fun getEndpoint(ctx: Context): Unit =
+    override fun updateEndpoint(ctx: Context) {
         Utils.query {
+            val session = ctx.attribute<User>("session")!!
             val resource = ctx.getResource()
-            ctx.json(resource.toJson(showAll = true))
-        }
-
-    fun updateEndpoint(ctx: Context): Unit =
-        Utils.query {
-            val resource = ctx.getResource()
+            if (session != resource && !(session.role >= UserRole.MODERATOR && session.role > resource.role)) {
+                throw UnauthorizedResponse()
+            }
             val input = ctx.getInput()
             val exists = User.find {
                 UserTable.usernameCol eq input.username
@@ -95,7 +89,7 @@ object UserApiRouter : Logging {
             if (exists != null && exists != resource) {
                 throw ConflictResponse(message = "User already exists")
             }
-            resource.imageUrl = input.imageUrl
+            resource.image = input.image
             input.readBooks.forEach {
                 val book = Book.findById(id = it.bookId)
                     ?: throw NotFoundResponse(message = "Unable to find Book: `${it.bookId}`")
@@ -118,84 +112,20 @@ object UserApiRouter : Logging {
 
             ctx.json(resource.toJson(showAll = true))
         }
+    }
 
-    fun deleteEndpoint(ctx: Context): Unit =
+    override fun deleteEndpoint(ctx: Context) {
         Utils.query {
+            val session = ctx.attribute<User>("session")!!
             val resource = ctx.getResource()
+            if (session != resource && !(session.role >= UserRole.MODERATOR && session.role > resource.role)) {
+                throw UnauthorizedResponse()
+            }
             resource.readBooks.forEach {
                 it.delete()
             }
             resource.delete()
             ctx.status(HttpStatus.NO_CONTENT)
         }
-
-    private fun Context.getReadInput(): ReadBookInput =
-        this.bodyValidator<ReadBookInput>()
-            .check({ it.bookId > 0 }, error = "BookId must be greater than 0")
-            .get()
-
-    fun addReadBook(ctx: Context): Unit =
-        Utils.query {
-            val resource = ctx.getResource()
-            val input = ctx.getReadInput()
-            val book = Book.findById(id = input.bookId)
-                ?: throw NotFoundResponse(message = "Unable to find Book: `${input.bookId}`")
-            val readBook = ReadBook.find {
-                (ReadBookTable.bookCol eq book.id) and (ReadBookTable.userCol eq resource.id)
-            }.firstOrNull() ?: ReadBook.new {
-                this.book = book
-                this.user = resource
-            }
-            readBook.readDate = input.readDate
-
-            ctx.json(resource.toJson(showAll = true))
-        }
-
-    private fun Context.getIdValue(): IdValue =
-        this.bodyValidator<IdValue>()
-            .check({ it.id > 0 }, error = "Id must be greater than 0")
-            .get()
-
-    fun removeReadBook(ctx: Context): Unit =
-        Utils.query {
-            val resource = ctx.getResource()
-            val input = ctx.getIdValue()
-            val book = Book.findById(id = input.id)
-                ?: throw NotFoundResponse(message = "Unable to find Book: `${input.id}`")
-            val exists = ReadBook.find {
-                (ReadBookTable.bookCol eq book.id) and (ReadBookTable.userCol eq resource.id)
-            }.firstOrNull() ?: throw BadRequestResponse(message = "User has not read this Book")
-            exists.delete()
-
-            ctx.json(resource.toJson(showAll = true))
-        }
-
-    fun addWishedBook(ctx: Context): Unit =
-        Utils.query {
-            val resource = ctx.getResource()
-            val input = ctx.getIdValue()
-            val book = Book.findById(id = input.id)
-                ?: throw NotFoundResponse(message = "Unable to find Book: `${input.id}`")
-            val temp = resource.wishedBooks.toMutableSet()
-            temp.add(book)
-            resource.wishedBooks = SizedCollection(temp)
-
-            ctx.json(resource.toJson(showAll = true))
-        }
-
-    fun removeWishedBook(ctx: Context): Unit =
-        Utils.query {
-            val resource = ctx.getResource()
-            val input = ctx.getIdValue()
-            val book = Book.findById(id = input.id)
-                ?: throw NotFoundResponse(message = "Unable to find Book: `${input.id}`")
-            if (!resource.wishedBooks.contains(book)) {
-                throw NotFoundResponse(message = "User has not wished for this Book")
-            }
-            val temp = resource.wishedBooks.toMutableSet()
-            temp.remove(book)
-            resource.wishedBooks = SizedCollection(temp)
-
-            ctx.json(resource.toJson(showAll = true))
-        }
+    }
 }
